requirements.txt
# Natural Language Processing
langchain-core>=0.3.15
langchain-community>=0.3.5
langchain-groq>=0.2.1
langchain-huggingface>=0.1.2
chromadb>=0.5.18  # Updated to latest available version
sentence-transformers
huggingface_hub
transformers
docx2txt
unstructured
unstructured[pdf]
rank-bm25
# Utilities
click
requests

# UI Frameworks
chainlit

# Excel File Manipulation
openpyxl

# Environment and Config
python-dotenv
rank_bm25